fix: label 삭제
라벨 인풋에 다시 병합
